Add 'send_message_packet' event
Added a better way to send messages.
This allows user scripts to easily send a message to multiple users and also add attachments (to be implemented).
There are also some bug fixes.
Fixed `send_message()` function not actually sending messages to other users.
Fixed messages that not meant to be to be sent to the local user being sent to the local user.